# doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Audio Files Hosting API Specs",
    "version": "1.0"
  },
  "paths": {
...
    "/api/audio-files": {
      "get": {
        "summary": "Get a list of audio files",
        "description": "Retrieve a list of audio files for the authenticated user.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "The page number to retrieve.",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "The number of audio files to retrieve per page.",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of audio files.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "description": {
                            "type": "string",
                            "example": "Sample audio file"
                          },
                          "category": {
                            "type": "string",
                            "example": "music"
                          },
                          "mimeType": {
                            "type": "string",
                            "example": "audio/mp3"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-10-01T12:00:00Z"
                          }
                        },
                        "required": [
                          "id",
                          "description",
                          "category",
                          "mimeType",
                          "createdAt"
                        ]
                      }
                    },
                    "totalCount": {
                      "type": "integer",
                      "example": 100
                    }
                  },
                  "required": [
                    "items",
                    "totalCount"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Failed to fetch audio files.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch audio files"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Session expired"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload a new audio file",
        "description": "Upload a new audio file for the authenticated user.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "A description of the audio file.",
                    "example": "Sample audio file"
                  },
                  "category": {
                    "type": "string",
                    "description": "The category of the audio file.",
                    "example": "music"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The audio file to upload."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio file uploaded successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Audio file uploaded successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to upload audio file.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "string",
                          "example": "Failed to upload audio file"
                        },
                        "description": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Description must be at least 3 characters long."
                          }
                        },
                        "file": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "File type must be specified."
                          }
                        },
                        "category": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "example": "Category must be one of the categories in the dropdown."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "string",
                          "example": "Session expired"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "User login",
        "description": "Create a new session for the user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "testuser"
                  },
                  "password": {
                    "type": "string",
                    "example": "testpassword"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login successful"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid username or password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid username or password"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}